chore: split out cdp wallet and cdp api actions (#204)
